Nimiro_AI/requirements.txt
# africastalking==2.0.2
# Django==6.0.1
# Flask==3.1.2
# google-genai==1.60.0
# google-generativeai==0.8.6
# langchain-google-genai==4.2.0
# python-dotenv==1.2.1
# requests==2.32.5
# requests-toolbelt==1.0.0
# whitenoise==6.11.0

Django
python-dotenv
google-generativeai
africastalking
Flask
google-genai